move a bunch of stuff to plz.py
